Do not import from client folder in smoke tests
