+++ package.json
@@ -455,7 +455,7 @@
                 },
                 "python.defaultInterpreterPath": {
                     "default": "python",
-                    "description": "Path to default Python to use when extension loads up for the first time, no longer used once an interpreter is selected for a workspace. See https://aka.ms/AAfekmf to understand when this is used.",
+                    "description": "Path to default Python to use when extension loads up for the first time, no longer used once an interpreter is selected for the workspace. See https://aka.ms/AAfekmf to understand when this is used.",
                     "scope": "machine-overridable",
                     "type": "string"
                 },
